Basic Login Prototype
A prototype of login functionality. When a user is logged into their
account the sign in and register forms are hidden. While the user is
logged on their email shows up as well as a sign out button. Right now
their is still no verification of account credentials from the database.